fix: challenge user - write post

+++ src/main/java/com/posmosalimos/geulgwi/entity/ChallengeAdmin.java
@@ -14,7 +14,7 @@ public class ChallengeAdmin {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(updatable = false)
-    private int challengeAdminSeq;
+    private Long challengeAdminSeq;
     private String keyword1;
     private String keyword2;
     private String keyword3;
@@ -22,7 +22,7 @@ public class ChallengeAdmin {
     private List<ChallengeUser> challengeUsers;
 
     @Builder
-    public ChallengeAdmin(String[] keyword, int keyword_seq) {
+    public ChallengeAdmin(String[] keyword, Long keyword_seq) {
         this.challengeAdminSeq = keyword_seq;
         this.keyword1 = keyword[0];
         this.keyword2 = keyword[1];
